Bug fixes, add rendering and uploading files, code improvements
This commit has all this shit because one time I started working on
things and I just didn't stop. Basically, it fixes some bugs, it adds
the abilit to render files through render.go (but I realise as I'm
writing this commit that there's no button to push to bring you to the
render page, oops), adds the ability to upload files, and other stuff
such as shortening some of the code. Good night